+++ src/main/java/com/Gintaras/tcgtrading/card_service/business/service/impl/CardServiceImpl.java
@@ -5,9 +5,11 @@ import com.Gintaras.tcgtrading.card_service.business.repository.CardRepository;
 import com.Gintaras.tcgtrading.card_service.business.repository.DAO.CardDAO;
 import com.Gintaras.tcgtrading.card_service.business.service.CardService;
 import com.Gintaras.tcgtrading.card_service.model.Card;
-import lombok.extern.log4j.Log4j2;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import lombok.extern.log4j.Log4j2;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,36 +20,60 @@ import java.util.stream.Collectors;
 public class CardServiceImpl implements CardService {
 
     @Autowired
-    CardRepository cardRepository;
+    private CardRepository cardRepository;
 
     @Autowired
-    CardMapStruct cardMapper;
+    private CardMapStruct cardMapper;
 
     @Override
-   public Card saveCard(Card card) {
+    public ResponseEntity<Card> saveCard(Card card) {
+        if (cardRepository.findById(card.getCardRarityId()).isEmpty()) {
+            log.warn("Card rarity with id {} does not exist", card.getCardRarityId());
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST); // Return 400 if rarity doesn't exist
+        }
+
         CardDAO cardDAO = cardRepository.save(cardMapper.CardToCardDAO(card));
         log.info("New Card is saved: {}", card);
-        return cardMapper.CardDAOToCard(cardDAO);
+
+        return new ResponseEntity<>(cardMapper.CardDAOToCard(cardDAO), HttpStatus.CREATED);
     }
 
     @Override
-    public Optional<Card> getCardById(String id){
+    public ResponseEntity<Card> getCardById(String id) {
         Optional<Card> card = cardRepository.findById(id).map(cardMapper::CardDAOToCard);
-        log.info("Card with id {} is {}", id, card.isPresent() ? card.get() : "not found");
-        return card;
+
+        if (card.isPresent()) {
+            log.info("Card with id {} found: {}", id, card.get());
+            return new ResponseEntity<>(card.get(), HttpStatus.OK);
+        } else {
+            log.info("Card with id {} is not found", id);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @Override
-    public void deleteCardById(String id){
-        cardRepository.deleteById(id);
-        log.info("Card  with id {} has been deleted", id);
+    public ResponseEntity<Void> deleteCardById(String id) {
+        if (cardRepository.existsById(id)) {
+            cardRepository.deleteById(id);
+            log.info("Card with id {} has been deleted", id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            log.warn("Card with id {} does not exist", id);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @Override
-    public List<Card> getCardList (){
+    public ResponseEntity<List<Card>> getCardList() {
         List<CardDAO> cardList = cardRepository.findAll();
-        log.info("Get card list. Size is: {}", cardList::size);
-        return cardList.stream().map(cardMapper::CardDAOToCard).collect(Collectors.toList());
+        log.info("Get card list. Size is: {}", cardList.size());
 
+        List<Card> cards = cardList.stream().map(cardMapper::CardDAOToCard).collect(Collectors.toList());
+
+        if (!cards.isEmpty()) {
+            return new ResponseEntity<>(cards, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 }
